refactor: removed visualize activity and renamed create activity to manage
