+++ app/src/main/java/com/example/projectmobile/ui/admin/cars/AdminCarsActivity.kt
@@ -14,8 +14,8 @@ import com.example.projectmobile.api.types.Car
 import com.example.projectmobile.databinding.ActivityAdminCarsBinding
 import com.example.projectmobile.databinding.ModalLayoutBinding
 import com.example.projectmobile.ui.admin.cars.adapter.AdminCarsAdapter
-import com.example.projectmobile.ui.admin.cars.car.CreateCarActivity
-import com.example.projectmobile.ui.admin.cars.car.VisualizeCarActivity
+import com.example.projectmobile.ui.admin.cars.clicklistener.CarClickListener
+import com.example.projectmobile.ui.admin.cars.manager.ManageCarActivity
 import com.example.projectmobile.util.UserPreferencesManager
 import java.io.IOException
 
@@ -24,10 +24,14 @@ class AdminCarsActivity : AppCompatActivity(), CarClickListener {
     private val binding get() = _binding!!
     private val adapter = AdminCarsAdapter(this)
 
+    private lateinit var preferencesManager: UserPreferencesManager
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         _binding = ActivityAdminCarsBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        preferencesManager = UserPreferencesManager(this)
 
         supportActionBar?.hide()
 
@@ -42,7 +46,7 @@ class AdminCarsActivity : AppCompatActivity(), CarClickListener {
         }
 
         binding.addButton.setOnClickListener {
-            startActivity(Intent(this, CreateCarActivity::class.java))
+            startActivity(Intent(this, ManageCarActivity::class.java))
         }
 
         handleSearch()
@@ -54,11 +58,10 @@ class AdminCarsActivity : AppCompatActivity(), CarClickListener {
 
     private fun handleSearch() {
         val search = binding.editResearch.text.toString()
-        val preferencesManager = UserPreferencesManager(this)
-        getVehiclesData(preferencesManager, search)
+        getVehiclesData(search)
     }
 
-    private fun getVehiclesData(preferencesManager: UserPreferencesManager, search: String) {
+    private fun getVehiclesData(search: String) {
         loading()
         val apiService = APIService(preferencesManager.getToken())
         val url = "/vehicle/list?page=1&size=100&sort=ASC&search=$search"
@@ -125,7 +128,8 @@ class AdminCarsActivity : AppCompatActivity(), CarClickListener {
     }
 
     override fun onEditCarClick(car: Car) {
-        startActivity(Intent(this@AdminCarsActivity, VisualizeCarActivity::class.java))
+        preferencesManager.saveSelectedCar(car)
+        startActivity(Intent(this@AdminCarsActivity, ManageCarActivity::class.java))
     }
 
     override fun onDeleteCarClick(car: Car) {
@@ -152,7 +156,6 @@ class AdminCarsActivity : AppCompatActivity(), CarClickListener {
 
     private fun deleteItem(car: Car) {
         loading()
-        val preferencesManager = UserPreferencesManager(this)
         val apiService = APIService(preferencesManager.getToken())
         val url = "/vehicle?id=${car.id}"
