+++ app/src/main/java/com/example/projectmobile/ui/admin/reservations/manager/ManageReservationActivity.kt
@@ -1,4 +1,4 @@
-package com.example.projectmobile.ui.admin.cars.car
+package com.example.projectmobile.ui.admin.reservations.manager
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -8,19 +8,19 @@ import android.widget.Toast
 import com.example.projectmobile.api.callback.APICallback
 import com.example.projectmobile.api.service.APIService
 import com.example.projectmobile.api.types.APIResponse
-import com.example.projectmobile.databinding.ActivityCreateCarBinding
+import com.example.projectmobile.databinding.ActivityCreateReservationBinding
 import com.example.projectmobile.util.UserPreferencesManager
 import java.io.IOException
 
-class CreateCarActivity : AppCompatActivity() {
-    private var _binding: ActivityCreateCarBinding? = null
+class ManageReservationActivity : AppCompatActivity() {
+    private var _binding: ActivityCreateReservationBinding? = null
     private val binding get() = _binding!!
 
     private lateinit var preferencesManager: UserPreferencesManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        _binding = ActivityCreateCarBinding.inflate(layoutInflater)
+        _binding = ActivityCreateReservationBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         preferencesManager = UserPreferencesManager(this)
@@ -32,47 +32,45 @@ class CreateCarActivity : AppCompatActivity() {
         }
 
         binding.registerButton.setOnClickListener {
-            val brand = binding.editBrand.text.toString()
-            val model = binding.editModel.text.toString()
-            val color = binding.editColor.text.toString()
-            val plate = binding.editPlate.text.toString()
-            val price = binding.editPrice.text.toString()
+            val userId = binding.userId.text.toString()
+            val vehicleId = binding.vehicleId.text.toString()
+            val pickup = binding.pickup.text.toString()
+            val devolution = binding.devolution.text.toString()
 
-            if (validateFields(brand, model, color, plate, price)) {
+            if (validateFields(userId, vehicleId, pickup, devolution)) {
                 sendDataToServer()
             }
         }
     }
 
+    override fun onDestroy() {
+        super.onDestroy()
+        preferencesManager.removeSelectedReservation()
+    }
+
     private fun validateFields(
-        brand: String,
-        model: String,
-        color: String,
-        plate: String,
-        price: String,
+        userId: String,
+        vehicleId: String,
+        pickup: String,
+        devolution: String,
     ): Boolean {
-        if (TextUtils.isEmpty(brand)) {
-            binding.editBrand.error = "Campo obrigatório"
+        if (TextUtils.isEmpty(userId)) {
+            binding.userId.error = "Campo obrigatório"
             return false
         }
 
-        if (TextUtils.isEmpty(model)) {
-            binding.editModel.error = "Campo obrigatório"
+        if (TextUtils.isEmpty(vehicleId)) {
+            binding.vehicleId.error = "Campo obrigatório"
             return false
         }
 
-        if (TextUtils.isEmpty(color)) {
-            binding.editColor.error = "Campo obrigatório"
+        if (TextUtils.isEmpty(pickup)) {
+            binding.pickup.error = "Campo obrigatório"
             return false
         }
 
-        if (TextUtils.isEmpty(plate)) {
-            binding.editPlate.error = "Campo obrigatório"
-            return false
-        }
-
-        if (TextUtils.isEmpty(price)) {
-            binding.editPrice.error = "Campo obrigatório"
+        if (TextUtils.isEmpty(devolution)) {
+            binding.devolution.error = "Campo obrigatório"
             return false
         }
 
@@ -82,7 +80,7 @@ class CreateCarActivity : AppCompatActivity() {
     private fun sendDataToServer() {
         loading()
         val apiService = APIService(preferencesManager.getToken())
-        val url = "/vehicle/register"
+        val url = "/reservation/register"
 
         val requestData = getRequestData()
 
@@ -93,7 +91,7 @@ class CreateCarActivity : AppCompatActivity() {
 
                     runOnUiThread {
                         Toast.makeText(
-                            this@CreateCarActivity,
+                            this@ManageReservationActivity,
                             message,
                             Toast.LENGTH_SHORT
                         ).show()
@@ -106,7 +104,7 @@ class CreateCarActivity : AppCompatActivity() {
                     runOnUiThread {
                         loaded()
                         Toast.makeText(
-                            this@CreateCarActivity,
+                            this@ManageReservationActivity,
                             errorCode,
                             Toast.LENGTH_SHORT
                         ).show()
@@ -118,7 +116,7 @@ class CreateCarActivity : AppCompatActivity() {
                 runOnUiThread {
                     loaded()
                     Toast.makeText(
-                        this@CreateCarActivity,
+                        this@ManageReservationActivity,
                         error.message,
                         Toast.LENGTH_SHORT
                     ).show()
@@ -128,28 +126,24 @@ class CreateCarActivity : AppCompatActivity() {
     }
 
     private fun getRequestData(): String {
-        val brand = binding.editBrand.text.toString()
-        val model = binding.editModel.text.toString()
-        val color = binding.editColor.text.toString()
-        val plate = binding.editPlate.text.toString()
-        val price = binding.editPrice.text.toString()
+        val userId = binding.userId.text.toString()
+        val vehicleId = binding.vehicleId.text.toString()
+        val pickup = binding.pickup.text.toString()
+        val devolution = binding.devolution.text.toString()
 
         return """
-            "brand": "$brand",
-            "model": "$model",
-            "color": "$color",
-            "plate": "$plate",
-            "value": "$price",
+            "userId": "$userId",
+            "vehicleId": "$vehicleId",
+            "pickup": "$pickup",
+            "devolution": "$devolution",
         """.trimIndent()
     }
 
     private fun loading() {
         binding.titleTextView.visibility = View.GONE
-        binding.editBrand.visibility = View.GONE
-        binding.editModel.visibility = View.GONE
-        binding.editColor.visibility = View.GONE
-        binding.editPlate.visibility = View.GONE
-        binding.editPrice.visibility = View.GONE
+        binding.idLayout.visibility = View.GONE
+        binding.pickup.visibility = View.GONE
+        binding.devolution.visibility = View.GONE
         binding.registerButton.visibility = View.GONE
 
         binding.progressBar.visibility = View.VISIBLE
@@ -157,11 +151,9 @@ class CreateCarActivity : AppCompatActivity() {
 
     private fun loaded() {
         binding.titleTextView.visibility = View.VISIBLE
-        binding.editBrand.visibility = View.VISIBLE
-        binding.editModel.visibility = View.VISIBLE
-        binding.editColor.visibility = View.VISIBLE
-        binding.editPlate.visibility = View.VISIBLE
-        binding.editPrice.visibility = View.VISIBLE
+        binding.idLayout.visibility = View.VISIBLE
+        binding.pickup.visibility = View.VISIBLE
+        binding.devolution.visibility = View.VISIBLE
         binding.registerButton.visibility = View.VISIBLE
 
         binding.progressBar.visibility = View.GONE
