+++ app/src/main/java/com/example/projectmobile/ui/admin/reservations/viewholder/AdminReservationViewHolder.kt
@@ -3,7 +3,7 @@ package com.example.projectmobile.ui.admin.reservations.viewholder
 import androidx.recyclerview.widget.RecyclerView
 import com.example.projectmobile.api.types.Reservation
 import com.example.projectmobile.databinding.RowAdminItemBinding
-import com.example.projectmobile.ui.admin.reservations.ReservationClickListener
+import com.example.projectmobile.ui.admin.reservations.clicklistener.ReservationClickListener
 import java.text.SimpleDateFormat
 import java.util.*
 
@@ -22,12 +22,12 @@ class AdminReservationViewHolder(private val bind: RowAdminItemBinding, private 
         }
     }
 
-    private fun dateFormatter(dataString: String): String {
+    private fun dateFormatter(dataString: String?): String {
         val entryFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", Locale.getDefault())
         val exitFormat = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
 
         if (dataString != null) {
-            return exitFormat.format(entryFormat.parse(dataString))
+            return exitFormat.format(entryFormat.parse(dataString)!!)
         }
 
         return "?/?/?"
