+++ app/src/main/java/com/example/projectmobile/ui/admin/users/clicklistener/UserClickListener.kt
@@ -1,4 +1,4 @@
-package com.example.projectmobile.ui.admin.users
+package com.example.projectmobile.ui.admin.users.clicklistener
 
 import com.example.projectmobile.api.types.User
