+++ app/src/main/java/com/example/projectmobile/ui/admin/users/viewholder/AdminUserViewHolder.kt
@@ -1,10 +1,9 @@
 package com.example.projectmobile.ui.admin.users.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
-import com.example.projectmobile.R
 import com.example.projectmobile.api.types.User
 import com.example.projectmobile.databinding.RowAdminItemBinding
-import com.example.projectmobile.ui.admin.users.UserClickListener
+import com.example.projectmobile.ui.admin.users.clicklistener.UserClickListener
 
 class AdminUserViewHolder(private val bind: RowAdminItemBinding, private val userClickListener: UserClickListener): RecyclerView.ViewHolder(bind.root) {
     fun bind(user: User) {
