+++ app/src/main/res/layout/activity_create_car.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.admin.cars.car.CreateCarActivity">
+    tools:context=".ui.admin.cars.manager.ManageCarActivity">
 
     <ImageButton
         android:id="@+id/returnButton"
